Bug fix: There was two voting.services.ts files other was in schemas folder. Deleted the wrong one and added a method from wrong to correct service.
Added: getVoting method
