+++ src/voting/voting.service.ts
@@ -1,18 +1,22 @@
 import { Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
-import { Voting } from "./voting.schema";
 import { Model } from "mongoose";
 import BasicService from "../common/service/basicService/BasicService";
 import { CreateVotingDto } from "./dto/createVoting.dto";
 import { ItemVoteChoice } from "./enum/choiceType.enum";
 import VotingNotifier from "./voting.notifier";
 import { StartItemVotingParams } from "./type/startItemVoting.type";
+import { Voting } from "./schemas/voting.schema";
+import { VotingDto } from "./dto/voting.dto";
+import ServiceError from "../common/service/basicService/ServiceError";
+import { PlayerService } from "../player/player.service";
 
 @Injectable()
 export class VotingService {
 	constructor(
 		@InjectModel(Voting.name) public readonly model: Model<Voting>,
-		private readonly notifier: VotingNotifier
+		private readonly notifier: VotingNotifier,
+		private readonly playerService: PlayerService
 	) {
 		this.basicService = new BasicService(model);
 	}
@@ -26,7 +30,7 @@ export class VotingService {
 	 * @returns A promise that resolves to the created voting entity.
 	 */
 	async createOne(voting: CreateVotingDto) {
-		return this.basicService.createOne<CreateVotingDto, Voting>(voting);
+		return this.basicService.createOne<CreateVotingDto, VotingDto>(voting);
 	}
 
 	/**
@@ -41,11 +45,13 @@ export class VotingService {
 	 *
 	 * @throws - Throws an error if validation fails or if there are errors creating the voting.
 	 */
-	async startItemVoting(params: StartItemVotingParams) {
+	async startItemVoting(
+		params: StartItemVotingParams
+	): Promise<[VotingDto, ServiceError[]]> {
 		const { playerId, itemId, clanId, type } = params;
 
 		const newVoting: CreateVotingDto = {
-			organizer_id: playerId,
+			organizer: { player_id: playerId, clan_id: clanId },
 			type: type,
 			entity_id: itemId,
 		};
@@ -55,12 +61,51 @@ export class VotingService {
 			choice: ItemVoteChoice.YES,
 		};
 
-		newVoting.player_ids = [playerId];
 		newVoting.votes = [newVote];
 
 		const [voting, errors] = await this.createOne(newVoting);
-		if (errors) throw errors;
+		if (errors) return [null, errors];
 
 		this.notifier.newVoting(clanId, voting);
+		return [voting, null];
+	}
+
+	/**
+	 * Checks if the voting has been successful.
+	 *
+	 * @param voting - The voting data to check.
+	 * @returns A boolean indicating whether the voting has been successful.
+	 */
+	async checkVotingSuccess(voting: VotingDto) {
+		const yesVotes = voting.votes.filter(
+			(vote) => vote.choice === ItemVoteChoice.YES
+		).length;
+		const totalVotes = voting.votes.length;
+		const yesPercentage = (yesVotes / totalVotes) * 100;
+
+		return yesPercentage >= voting.minPercentage;
+	}
+
+	/**
+	 * Reads a voting from DB by id and checks if user has permission to view it.
+	 *
+	 * @param votingId - The ID of the voting to fetch.
+	 * @param playerId - The ID of the player.
+	 *
+	 * @returns The voting entry if the user has permission to view it, otherwise null.
+	 * @throws Will throw if there is an error reading from DB.
+	 */
+	async getVoting(votingId: string, playerId: string) {
+		const [voting, errors] = await this.basicService.readOneById<VotingDto>(
+			votingId
+		);
+		if (errors) throw errors;
+
+		if (!voting.organizer.clan_id) return voting;
+
+		const clanId = await this.playerService.getPlayerClanId(playerId);
+		if (clanId === voting.organizer.clan_id) return voting;
+
+		return null;
 	}
 }
